chore: improve startup time
